requirements.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
atari-py==0.2.6
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.2
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
DataProperty==0.43.1
decorator==4.4.0
dm-control==0.0.0
dm-env==1.1
enum34==1.1.6
future==0.17.1
gast==0.2.2
gin-config==0.2.1
glfw==1.8.3
google-pasta==0.1.7
grpcio==1.24.0
gym==0.14.0
h5py==2.10.0
hdfs3==0.3.1
idna==2.8
imageio==2.5.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.1
mbstrdecoder==0.8.1
msgfy==0.0.7
mujoco-py==2.0.2.5
networkx==2.4
numpy==1.17.2
nvidia-ml-py3==7.352.0
opencv-python==4.1.1.26
pandas==0.25.3
pathvalidate==0.29.0
Pillow==6.2.0
protobuf==3.9.2
pybullet==2.5.6
-e git+https://github.com/benelot/pybullet-gym.git@55eaa0defca7f4ae382963885a334c952133829d#egg=pybulletgym
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.4.2
pytablewriter==0.46.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
tabledata==0.9.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
typepy==0.6.0
Werkzeug==0.16.0
wrapt==1.11.2
xvfbwrapper==0.2.9
